+++ src/compounds/navigation-dropdown/CHANGELOG.md
@@ -3,6 +3,19 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.1](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.navigation-dropdown@0.1.0...@uswitch/trustyle.navigation-dropdown@0.1.1) (2021-09-27)
+
+
+### Bug Fixes
+
+* change React.createElement to jsx ([d495e14](https://github.com/uswitch/trustyle/commit/d495e14))
+* config links in story ([6dc74e0](https://github.com/uswitch/trustyle/commit/6dc74e0))
+* style tweaks and bug fixes ([d52310c](https://github.com/uswitch/trustyle/commit/d52310c))
+
+
+
+
+
 # 0.1.0 (2021-09-23)